Youtube_Video_Summarize/requirements.txt
streamlit
yt-dlp
pydub
openai-whisper
openai
youtube-transcript-api